refactor: migrate project scanning logic to a dedicated adapter and remove deprecated parsers
