+++ packages/cli/src/adapters/index.ts
@@ -3,3 +3,4 @@ export * from "./logger.adapter";
 export * from "./node-workspace";
 export * from "./prompt.adapter";
 export * from "./reporters";
+export * from "./node-project-scanner";
