+++ packages/cli/src/cli/commands/init.ts
@@ -3,14 +3,15 @@ import pc from "picocolors";
 import { code } from "@/cli/ui";
 import {
 	type CollectedState,
-	checkTsConfig,
 	createPlan,
-	detectFramework,
-	detectPackageManager,
 	Executor,
-	suggestDefaultEnvPath,
 } from "@/features/scaffold";
-import type { LoggerPort, PromptPort, WorkspacePort } from "@/shared/ports";
+import type {
+	LoggerPort,
+	PromptPort,
+	WorkspacePort,
+	ProjectScannerPort,
+} from "@/shared/ports";
 
 /**
  * Input parameters for the 'init' command.
@@ -29,6 +30,7 @@ export class InitUseCase {
 		private readonly logger: LoggerPort,
 		private readonly workspace: WorkspacePort,
 		private readonly prompt: PromptPort,
+		private readonly scanner: ProjectScannerPort,
 	) {}
 
 	async execute(input: InitInput) {
@@ -53,7 +55,7 @@ export class InitUseCase {
 
 		try {
 			let shouldUpdateTsConfig = false;
-			const tsConfig = await checkTsConfig();
+			const tsConfig = await this.scanner.checkTsConfig();
 
 			if (tsConfig.status === "not_strict") {
 				if (isYes) {
@@ -77,16 +79,38 @@ export class InitUseCase {
 				}
 			}
 
-			const detectedFramework = await detectFramework(tsConfig.parsed);
-			const defaultEnvPath = await suggestDefaultEnvPath(
+			const detectedFramework = await this.scanner.detectFramework(
 				process.cwd(),
 				tsConfig.parsed,
 			);
+			const defaultEnvPath = await this.scanner.suggestDefaultEnvPath(
+				process.cwd(),
+				tsConfig.parsed,
+			);
+
+			const targetPath = path.resolve(process.cwd(), defaultEnvPath);
+			const envRes = await this.scanner.getEnvExampleKeys(
+				process.cwd(),
+				tsConfig.parsed,
+				targetPath,
+			);
+
+			let hasTypeFile = false;
+			if (detectedFramework === "vite" || detectedFramework === "bun") {
+				const typeFile = detectedFramework === "vite" ? "vite-env.d.ts" : "bun-env.d.ts";
+				const targetDir = path.dirname(targetPath);
+				const typeFilePath = path.join(targetDir, typeFile);
+				hasTypeFile = await this.workspace.exists(typeFilePath);
+			}
+
 			const options = await this.prompt.runWizard(
 				{
 					framework: detectedFramework,
 					defaultEnvPath,
 					tsConfig: tsConfig.parsed,
+					envKeys: envRes?.keys,
+					envKeysSource: envRes?.source,
+					hasTypeFile,
 				},
 				isYes,
 			);
@@ -114,14 +138,14 @@ export class InitUseCase {
 			}
 
 			// Handle existing env file prompt
-			const targetPath = path.resolve(process.cwd(), options.path);
+			const finalTargetPath = path.resolve(process.cwd(), options.path);
 
 			if (
-				(await this.workspace.exists(targetPath)) &&
+				(await this.workspace.exists(finalTargetPath)) &&
 				options.overwriteEnvSchemaFile === undefined
 			) {
 				const confirmOverwrite = await this.prompt.confirm(
-					`File ${code(path.basename(targetPath))} already exists. Overwrite?`,
+					`File ${code(path.basename(finalTargetPath))} already exists. Overwrite?`,
 					false,
 				);
 
@@ -133,11 +157,14 @@ export class InitUseCase {
 				options.overwriteEnvSchemaFile = confirmOverwrite;
 			}
 
-			const packageManager = await detectPackageManager(tsConfig.parsed);
+			const packageManager = await this.scanner.detectPackageManager(
+				process.cwd(),
+				tsConfig.parsed,
+			);
 
 			const existingFiles: string[] = [];
-			if (await this.workspace.exists(targetPath))
-				existingFiles.push(targetPath);
+			if (await this.workspace.exists(finalTargetPath))
+				existingFiles.push(finalTargetPath);
 
 			let typeFileName: string | undefined;
 			if (options.framework === "vite") {
@@ -147,7 +174,7 @@ export class InitUseCase {
 			}
 
 			if (typeFileName) {
-				const targetDir = path.dirname(targetPath);
+				const targetDir = path.dirname(finalTargetPath);
 				const typeFilePath = path.join(targetDir, typeFileName);
 				if (await this.workspace.exists(typeFilePath))
 					existingFiles.push(typeFilePath);
