+++ packages/cli/src/cli/composition.ts
@@ -1,4 +1,4 @@
-import { ClackPromptAdapter, NodeWorkspace } from "@/adapters";
+import { ClackPromptAdapter, NodeWorkspace, NodeProjectScannerAdapter } from "@/adapters";
 import { CLI } from "./cli";
 import { HelpUseCase, InitUseCase } from "./commands";
 
@@ -14,8 +14,9 @@ export function compose(argv: string[]) {
 	const logger = cli.logger; // CLI currently creates the logger, which is fine for now
 	const workspace = new NodeWorkspace(cli.isQuiet, logger.stdio);
 	const prompt = new ClackPromptAdapter();
+	const scanner = new NodeProjectScannerAdapter();
 
-	const initUseCase = new InitUseCase(logger, workspace, prompt);
+	const initUseCase = new InitUseCase(logger, workspace, prompt, scanner);
 	const helpUseCase = new HelpUseCase(logger);
 
 	return {
