+++ packages/cli/src/features/scaffold/index.ts
@@ -1,7 +1,5 @@
-export * from "./env-parser";
 export * from "./env-template";
 export * from "./executor";
 export * from "./plan";
 export * from "./planner";
 export * from "./scaffold";
-export * from "./tsconfig-parser";
