+++ packages/cli/src/shared/ports/index.ts
@@ -1,3 +1,4 @@
 export * from "./logger.port";
 export * from "./prompt.port";
 export * from "./workspace.port";
+export * from "./project-scanner.port";
